# gateway/openapi/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "payload/manager/services/v1/payload_admin_service.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/payloads": {
      "get": {
        "summary": "Returns a summary row for every payload known to the service.\nOptionally filtered by tier.",
        "operationId": "PayloadCatalogService_ListPayloads",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListPayloadsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "tierFilter",
            "description": "Optional tier filter; TIER_UNSPECIFIED means return all tiers.\n\n - TIER_UNSPECIFIED: Not set; treated as an error in most contexts.\n - TIER_GPU: GPU device memory (fastest, smallest, most volatile).\n - TIER_RAM: Host DRAM / shared-memory segment.\n - TIER_DISK: Local NVMe/SSD/HDD (durable, larger, slower than RAM).\n - TIER_OBJECT: Remote object storage (S3/GCS/Azure; durable, largest).",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "TIER_UNSPECIFIED",
              "TIER_GPU",
              "TIER_RAM",
              "TIER_DISK",
              "TIER_OBJECT"
            ],
            "default": "TIER_UNSPECIFIED"
          },
          {
            "name": "pageSize",
            "description": "Maximum number of payloads to return. 0 = server default (50); max 500.",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageToken",
            "description": "Opaque cursor from a previous ListPayloadsResponse.next_page_token.\nOmit or leave empty to start from the beginning.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "PayloadCatalogService"
        ]
      },
      "post": {
        "summary": "Allocate reserves placement but the payload is invisible to readers\nuntil CommitPayload succeeds.",
        "operationId": "PayloadCatalogService_AllocatePayload",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1AllocatePayloadResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Allocate reserves placement but payload is not visible\nuntil CommitPayload succeeds.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1AllocatePayloadRequest"
            }
          }
        ],
        "tags": [
          "PayloadCatalogService"
        ]
      }
    },
    "/v1/payloads/spill": {
      "post": {
        "summary": "Requests durable storage (usually disk/object storage).",
        "description": "BLOCKING policy waits until durable or failure.",
        "operationId": "PayloadCatalogService_Spill",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SpillResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SpillRequest"
            }
          }
        ],
        "tags": [
          "PayloadCatalogService"
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "SpillResponseResult": {
      "type": "object",
      "properties": {
        "id": {
          "$ref": "#/definitions/v1PayloadID"
        },
        "ok": {
          "type": "boolean"
        },
        "errorMessage": {
          "type": "string"
        },
        "payloadDescriptor": {
          "$ref": "#/definitions/v1PayloadDescriptor"
        }
      }
    },
    "v1AllocatePayloadRequest": {
      "type": "object",
      "properties": {
        "sizeBytes": {
          "type": "string",
          "format": "uint64"
        },
        "preferredTier": {
          "$ref": "#/definitions/v1Tier"
        },
        "ttlMs": {
          "type": "string",
          "format": "uint64"
        },
        "noEvict": {
          "type": "boolean"
        },
        "evictionPolicy": {
          "$ref": "#/definitions/v1EvictionPolicy"
        }
      },
      "description": "Allocate reserves placement but payload is not visible\nuntil CommitPayload succeeds."
    },
    "v1AllocatePayloadResponse": {
      "type": "object",
      "properties": {
        "payloadDescriptor": {
          "$ref": "#/definitions/v1PayloadDescriptor"
        }
      }
    },
...
    "v1DiskLocation": {
      "type": "object",
      "properties": {
        "path": {
          "type": "string"
        },
        "offsetBytes": {
          "type": "string",
          "format": "uint64"
        },
        "lengthBytes": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "v1EvictionPolicy": {
      "type": "object",
      "properties": {
        "priority": {
          "$ref": "#/definitions/v1EvictionPriority",
          "title": "Relative eviction importance"
        },
        "requireDurable": {
          "type": "boolean",
          "title": "Require durable copy before removing from volatile tier"
        },
        "spillTarget": {
          "$ref": "#/definitions/v1Tier",
          "title": "Preferred durability tier (usually DISK)"
        },
        "minResidencyTier": {
          "$ref": "#/definitions/v1Tier",
          "title": "Attempt to keep payload at or above this tier"
        }
      },
      "description": "Advisory eviction hints.\n\nThese are not strict guarantees \u2014 the system MAY violate them\nunder critical resource pressure."
    },
    "v1EvictionPriority": {
      "type": "string",
      "enum": [
        "EVICTION_PRIORITY_UNSPECIFIED",
        "EVICTION_PRIORITY_LOW",
        "EVICTION_PRIORITY_NORMAL",
        "EVICTION_PRIORITY_HIGH",
        "EVICTION_PRIORITY_NEVER"
      ],
      "default": "EVICTION_PRIORITY_UNSPECIFIED",
      "description": " - EVICTION_PRIORITY_UNSPECIFIED: Treated as NORMAL.\n - EVICTION_PRIORITY_LOW: Evict before NORMAL candidates.\n - EVICTION_PRIORITY_NORMAL: Default eviction weight.\n - EVICTION_PRIORITY_HIGH: Evict only after LOW and NORMAL are exhausted.\n - EVICTION_PRIORITY_NEVER: Never automatically evicted (implies no_evict=true)."
    },
...
    "v1GpuLocation": {
      "type": "object",
      "properties": {
        "deviceId": {
          "type": "integer",
          "format": "int64"
        },
        "ipcHandle": {
          "type": "string",
          "format": "byte",
          "title": "opaque CUDA IPC handle"
        },
        "lengthBytes": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
...
    "v1ListPayloadsResponse": {
      "type": "object",
      "properties": {
        "payloads": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1PayloadSummary"
          }
        },
        "nextPageToken": {
          "type": "string",
          "description": "Opaque cursor for the next page. Empty when this is the last page."
        },
        "totalCount": {
          "type": "integer",
          "format": "int32",
          "description": "Total number of payloads matching the filter (ignoring pagination)."
        }
      }
    },
...
    "v1PayloadDescriptor": {
      "type": "object",
      "properties": {
        "payloadId": {
          "$ref": "#/definitions/v1PayloadID"
        },
        "tier": {
          "$ref": "#/definitions/v1Tier"
        },
        "state": {
          "$ref": "#/definitions/v1PayloadState"
        },
        "gpu": {
          "$ref": "#/definitions/v1GpuLocation"
        },
        "ram": {
          "$ref": "#/definitions/v1RamLocation"
        },
        "disk": {
          "$ref": "#/definitions/v1DiskLocation"
        },
        "version": {
          "type": "string",
          "format": "uint64"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "expiresAt": {
          "type": "string",
          "format": "date-time"
        },
        "evictionPolicy": {
          "$ref": "#/definitions/v1EvictionPolicy",
          "title": "advisory eviction behavior only"
        }
      },
      "description": "Authoritative placement descriptor.\n\nversion increments whenever:\n- tier changes\n- location changes\n- lifecycle state changes\n\nClients MUST invalidate cached descriptors when version changes."
    },
    "v1PayloadID": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "format": "byte"
        }
      },
      "description": "Canonical payload identifier.\n\nMUST contain exactly 16 bytes (binary RFC4122 UUID).\nThe server rejects any other length.\n\nThis is intentionally NOT a string to avoid formatting ambiguity."
    },
...
    "v1PayloadState": {
      "type": "string",
      "enum": [
        "PAYLOAD_STATE_UNSPECIFIED",
        "PAYLOAD_STATE_ALLOCATED",
        "PAYLOAD_STATE_ACTIVE",
        "PAYLOAD_STATE_SPILLING",
        "PAYLOAD_STATE_DURABLE",
        "PAYLOAD_STATE_EVICTING",
        "PAYLOAD_STATE_DELETING",
        "PAYLOAD_STATE_EXPIRED",
        "PAYLOAD_STATE_DELETED"
      ],
      "default": "PAYLOAD_STATE_UNSPECIFIED",
      "description": "Lifecycle states are monotonic.\nClients MUST NOT assume a reversible transition.\n\n - PAYLOAD_STATE_UNSPECIFIED: Not set; treated as an error.\n - PAYLOAD_STATE_ALLOCATED: Reserved; writer has not yet committed data.\n - PAYLOAD_STATE_ACTIVE: Committed and readable by consumers.\n - PAYLOAD_STATE_SPILLING: Being moved to a lower tier; temporarily locked.\n - PAYLOAD_STATE_DURABLE: Confirmed written to a durable tier (DISK/OBJECT).\n - PAYLOAD_STATE_EVICTING: Being removed from its current tier under pressure.\n - PAYLOAD_STATE_DELETING: Deletion in progress.\n - PAYLOAD_STATE_EXPIRED: TTL elapsed; pending final deletion.\n - PAYLOAD_STATE_DELETED: Fully deleted; ID is no longer valid."
    },
    "v1PayloadSummary": {
      "type": "object",
      "properties": {
        "id": {
          "$ref": "#/definitions/v1PayloadID"
        },
        "tier": {
          "$ref": "#/definitions/v1Tier"
        },
        "state": {
          "$ref": "#/definitions/v1PayloadState"
        },
        "sizeBytes": {
          "type": "string",
          "format": "uint64"
        },
        "createdAtMs": {
          "type": "string",
          "format": "uint64",
          "title": "0 = unknown (legacy row)"
        },
        "expiresAtMs": {
          "type": "string",
          "format": "uint64",
          "title": "0 = no expiration"
        },
        "activeLeases": {
          "type": "integer",
          "format": "int64"
        }
      },
      "description": "Summary row returned by ListPayloads."
    },
...
    "v1RamLocation": {
      "type": "object",
      "properties": {
        "shmName": {
          "type": "string"
        },
        "slabId": {
          "type": "integer",
          "format": "int64"
        },
        "blockIndex": {
          "type": "string",
          "format": "uint64"
        },
        "lengthBytes": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "v1SpillPolicy": {
      "type": "string",
      "enum": [
        "SPILL_POLICY_UNSPECIFIED",
        "SPILL_POLICY_BEST_EFFORT",
        "SPILL_POLICY_BLOCKING"
      ],
      "default": "SPILL_POLICY_UNSPECIFIED",
      "description": "Spill policy controls durability operations.\n\n - SPILL_POLICY_BEST_EFFORT: Background durability\n - SPILL_POLICY_BLOCKING: Wait until durable or failure"
    },
    "v1SpillRequest": {
      "type": "object",
      "properties": {
        "ids": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1PayloadID"
          }
        },
        "policy": {
          "$ref": "#/definitions/v1SpillPolicy"
        },
        "waitForLeases": {
          "type": "boolean"
        },
        "fsync": {
          "type": "boolean"
        },
        "targetTier": {
          "$ref": "#/definitions/v1Tier",
          "description": "Optional override: spill to this tier instead of the per-payload default.\nMust be a tier below the payload's current tier."
        }
      }
    },
    "v1SpillResponse": {
      "type": "object",
      "properties": {
        "results": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/SpillResponseResult"
          }
        }
      }
    },
    "v1Tier": {
      "type": "string",
      "enum": [
        "TIER_UNSPECIFIED",
        "TIER_GPU",
        "TIER_RAM",
        "TIER_DISK",
        "TIER_OBJECT"
      ],
      "default": "TIER_UNSPECIFIED",
      "description": "Ordered from fastest \u2192 most durable.\nHigher tiers may not always be available.\n\n - TIER_UNSPECIFIED: Not set; treated as an error in most contexts.\n - TIER_GPU: GPU device memory (fastest, smallest, most volatile).\n - TIER_RAM: Host DRAM / shared-memory segment.\n - TIER_DISK: Local NVMe/SSD/HDD (durable, larger, slower than RAM).\n - TIER_OBJECT: Remote object storage (S3/GCS/Azure; durable, largest)."
    },
...
  }
}